Expose the transparent hugepages (THP) configuration so the `-XX:+UseTransparentHugePages` java setting works
Java is able to make use of hugepages to greatly reduce pagetable
lookups by allowing the process to map larger chunks of RAM at a time.
This is known to boost Minecraft performance on both the server and
client. The Linux kernel in particular offers transparent hugepages,
which allows userspace processes to specify whether they prefer
hugepages, without needing root access. In order for Java to detect
this, it needs to see the kernel-side switches in
`/sys/kernel/mm/transparent_hugepage/`. This change also allows access
to the manual hugepage configuration, for those who still use that.

Signed-off-by: lordpipe <lordpipe@protonmail.com>